AklatBayon/aklatbayon_api/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
mysqlclient>=2.2
Pillow>=10.0
